Time: 9 ms (21.31%), Space: 13.7 MB (45.41%) - LeetHub

+++ 0330-patching-array/0330-patching-array.cpp
@@ -3,7 +3,7 @@ class Solution
 public:
     int minPatches(vector<int>& nums, int n) 
     {
-        long long miss = 1;
+        long miss = 1;
         int ans = 0, i = 0;
         while (miss <= n) 
         {
